Conflicts:
	00-RELEASENOTES
	deps/hiredis/Makefile
	deps/hiredis/hiredis.c
	deps/hiredis/hiredis.h
	deps/hiredis/net.c
	redis.conf
	src/Makefile
	src/aof.c
	src/config.c
	src/db.c
	src/intset.c
	src/networking.c
	src/object.c
	src/rdb.c
	src/redis-benchmark.c
	src/redis-cli.c
	src/redis.c
	src/redis.h
	src/replication.c
	src/sort.c
	src/t_list.c
	src/t_string.c
	src/t_zset.c
	src/util.c
	src/version.h
	src/ziplist.c
	src/zmalloc.c
	src/zmalloc.h
	tests/integration/aof.tcl
	tests/integration/replication.tcl
	tests/test_helper.tcl
	tests/unit/slowlog.tcl
	tests/unit/type/list.tcl
	tests/unit/type/zset.tcl

+++ src/util.c
@@ -133,7 +133,7 @@ int stringmatchlen(const char *pattern, int patternLen,
 }
 
 int stringmatch(const char *pattern, const char *string, int nocase) {
-    return stringmatchlen(pattern,strlen(pattern),string,strlen(string),nocase);
+    return stringmatchlen(pattern,(int)strlen(pattern),string,(int)strlen(string),nocase);
 }
 
 /* Convert a string representing an amount of memory into the number of
@@ -172,7 +172,7 @@ long long memtoll(const char *p, int *err) {
         if (err) *err = 1;
         mul = 1;
     }
-    digits = u-p;
+    digits = (unsigned int)(u-p);
     if (digits >= sizeof(buf)) {
         if (err) *err = 1;
         return LLONG_MAX;
@@ -204,7 +204,7 @@ int ll2string(char *s, size_t len, long long value) {
     if (l+1 > len) l = len-1; /* Make sure it fits, including the nul term */
     memcpy(s,p,l);
     s[l] = '\0';
-    return l;
+    return (int)l;
 }
 
 /* Convert a string into a long long. Returns 1 if the string could be parsed
